# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/v1/auth/mfa/verify": {
      "post": {
        "operationId": "AuthController_verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/v1/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "XENAuth API",
    "description": "The official XENAuth API documentation",
    "version": "1.0",
    "contact": {
      "name": "Erzen Krasniqi",
      "url": "https://erzen.tk",
      "email": "erzenkrasniqi@matrics.io"
    },
    "termsOfService": "https://erzen.tk/terms",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
...
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "LOCAL server"
    },
    {
      "url": "https://api.erzen.xyz",
      "description": "DEV server"
    },
    {
      "url": "https://apis.erzen.xyz",
      "description": "PRODUCTION server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "name": {
            "type": "string",
            "description": "User full name"
          },
          "username": {
            "type": "string",
            "description": "User username"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string",
            "description": "User birthdate"
          },
          "language": {
            "type": "string",
            "description": "User preferred language"
          },
          "timezone": {
            "type": "string",
            "description": "User timezone"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "username",
          "birthdate",
          "language",
          "timezone"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MfaDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "code": {
            "type": "string",
            "description": "MFA code"
          }
        },
        "required": [
          "email",
          "password",
          "code"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          }
        },
        "required": [
          "email"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}